Restructured to avoid global variables for params
As I did before with the pyxtalviewer part, I changed the pyxtalmain
part so that all of the Tk variables used as parameters are now defined
as attributes of the pmw object, rather than global variables.  This
means that the GUI code generated by PAGE will have to be modified by
hand each time now.