Fixed apparent typo in qhull options
I hadn't thought much about this option before, but apparently the "J"
needs to be capitalized.  This "Joggles" the points by a small amount to
avoid some geometric singularity.  I only ran into problems with this
for a few larger images; it worked great with the incorrect, lowercase
"j" up until now.  I don't fully understand this, but I'll go with this
for now.