Made plotting bonds and dislocations faster
Previously, I had used nested for loops to compile the lists of line
segments for plotting the blue bonds and yellow dislocations.  Although
that was somewhat easier to read and understand, it was slow (together,
they added ~0.4 seconds for fairly large images), and not consistent
with the preferred style of numpy.